Extract MessageFilter to a top level test utility class
